+++ lib/td_auth/permissions/permission.ex
@@ -49,7 +49,8 @@ defmodule TdAuth.Permissions.Permission do
     view_deprecated_ingests: "view_deprecated_ingests",
     manage_confidential_structures: "manage_confidential_structures",
     manage_ingest_relations: "manage_ingest_relations",
-    view_data_structures_profile: "view_data_structures_profile"
+    view_data_structures_profile: "view_data_structures_profile",
+    view_quality_rule: "view_quality_rule"
   }
 
   schema "permissions" do