+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- [TD-1776] add permission view_quality_rule
+
 ## [3.1.0] 2019-07-08
 
 ### Changed